feat: add new eaglesong by Hamster, add eaglesong test
