+++ eth-contracts/contracts/test/TestEaglesong.sol
@@ -1,8 +1,22 @@
-pragma solidity ^0.5.10;
-import {EaglesongLib} from "../libraries/EaglesongLib.sol";
+pragma solidity ^0.7.4;
+import {Eaglesong} from "../Eaglesong.sol";
 
 contract TestEaglesong {
-    function ckbEaglesong(bytes memory data) public returns(bytes32) {
-        return EaglesongLib.EaglesongHash(data);
+    address public songAddr;
+
+    constructor(address _songAddr) {
+        songAddr = _songAddr;
+    }
+
+    function ckbEaglesong(bytes memory data) public returns (bytes32 result) {
+        bytes memory payload = abi.encodePacked(
+            hex"aa6d7de4",
+            data
+        );
+
+        (bool success, bytes memory returnData) = songAddr.staticcall(payload);
+        assembly {
+            result := mload(add(returnData, 0x20))
+        }
     }
 }
